requirements.txt
djangorestframework
django-cors-headers
django-filter
djoser
social-auth-app-django
django-markdownify
django-taggit
channels
django-celery-beat
django-celery-results
django-debug-toolbar
django-storages
django-ckeditor
django-crispy-forms
crispy-tailwind
